requirements.txt
librosa==0.11.0
matplotlib==3.8.4
noisereduce==3.0.3
numpy>=1.26.4
scikit_learn==1.3.2
scipy==1.15.3
tensorflow==2.19.0
umap_learn==0.5.7
